Extract `getBlockquoteRawText`
